chore: clear pre-existing golangci-lint issues
- internal/consensus/adaptor/identity.go: drop trailing blank line
  flagged by goimports.
- internal/cli/replay.go, internal/cli/replay_range.go: drop redundant
  int64() casts; the source fields are already int64.
- internal/tx/check/check_create.go: drop redundant uint32() casts on
  LsfHighFreeze / LsfLowFreeze, which are already typed as uint32.

No behavior change; just appeases unconvert and goimports so the lint
gate is clean again.

+++ internal/cli/replay.go
@@ -335,8 +335,8 @@ func executeReplayVerbose(state *StateFixture, env *EnvFixture, txs *TxsFixture,
 		return nil, nil, fmt.Errorf("parsing parent_hash: %w", err)
 	}
 
-	closeTime := time.Unix(protocol.RippleEpochUnix+int64(env.CloseTime), 0).UTC()
-	parentCloseTime := time.Unix(protocol.RippleEpochUnix+int64(env.ParentCloseTime), 0).UTC()
+	closeTime := time.Unix(protocol.RippleEpochUnix+env.CloseTime, 0).UTC()
+	parentCloseTime := time.Unix(protocol.RippleEpochUnix+env.ParentCloseTime, 0).UTC()
 
 	ledgerHeader := header.LedgerHeader{
 		LedgerIndex:         env.LedgerIndex,
